tests for hashing work! also got the time tests down

+++ Inc/hashFunc.h
@@ -11,7 +11,7 @@ enum types {
 class hashFunc {
     public:
         hashFunc(); // default
-        hashFunc(int aIn, int bIn, int pIn, int mIn); // type 1
+        hashFunc(double aIn, double bIn, int pIn, int mIn); // type 1
         hashFunc(int sIn, int sizeIn); // type 2
         hashFunc(const hashFunc &h); // copy
         hashFunc& operator=(const hashFunc &h); // copy operator
@@ -20,8 +20,8 @@ class hashFunc {
 
     private:
         // type 1 (prime number)
-        int a = -1;
-        int b = -1;
+        double a = -1;
+        double b = -1;
         int p = -1;
         int m = -1;
